Update WebSoccer.class.php
The new class "Model" is for all models, which defines the required basic functions, which is overwritten by the executing class depending on the desired change of the basic code. If further variables are filled with data in __construct, this part has been transferred to renderView(). The references to the instance classes are defined in the __construct and is now only defined in the class Model, which requires 16 KB less code.