Update jobs.php
- Fixed: Cron jobs page not see

The internal execution has not been working properly for a very long time, since the servers only have a script runtime of 30 seconds instead of 5 minutes. Therefore, the jobs must be carried out in the administration of the web space and according to https://github.com/ihofmann/open-websoccer/wiki/15.-Setup-Cron-jobs.